fix bug introduced by seqtype parsing
storing the seqtype in a tuple along with samplename and filepath resulted in the emission of null objects after chromosome splitting, because there were more output elements than elements in the seqtype list. This fixes the problem by using a .combine() call to assign the seqtype to each sample split individually.