don't fail when context doesn't occur in ref
When dealing with organisms that have many small contigs in their reference genome,
MethylExtract fails if a particular context is not present in the whole contig.

Specifically this happens when MethylExtract.pl tries to write coverage statistics to a log file,
it tries to divide by zero and then terminates with exit code 25.

see:
https://github.com/bioinfoUGR/methylextract/blob/6a5d2120abbbcd281c2f9acddce5f23ef674361b/MethylExtract.pl#L2823

This workaround adds 25 as a valid exit code and resumes the execution of the process block.